Breadcrumbs React port: fix & delete broken tests
 - fixes race conditions in tests where the global singleton
   `FauxtonAPI` was monkey patched later in other modules, but
   `fauxton/base.js` wasn't loaded
- deletes tests that messed with references / monkey patching and
   did not add much value
 - fixes a few restore calls
